% Generated by roxygen2 (4.0.2): do not edit by hand
\name{code.poly}
\alias{code.poly}
\title{Build polynomial predictor variables}
\usage{
code.poly(df = NULL, predictor = NULL, poly.order = NULL,
  orthogonal = TRUE, draw.poly = TRUE)
}
\arguments{
\item{df}{data frame, should not contain any variables called "predictor"}

\item{predictor}{string name of predictor variable}

\item{poly.order}{integer order of polynomial to be created}

\item{orthogonal}{logical value indicating whether polynomial should be orthogonal (default) or natural (aka raw)}

\item{draw.poly}{logical value indicating whether to create a graph showing tranformed polynomial predictor values, defaults to TRUE}
}
\value{
Returns a data frame containing the original data and at least two new columns: "predictor".Index, and a column for each order of the polynomial-transformed predictor
}
\description{
Takes a data frame, name of predictor variable, and polynomial order. Creates polynomial-transformed predictor variables, adds them to the data frame and returns the result. Original data frame is unchanged, remember to assign the result.
}
\details{
This function was originally written by Matt Winn and adapted and revised with his help.
}
\section{Contributors}{

Originally written by Matt Winn \url{http://www.mattwinn.com/tools/R_add_polynomials_to_df.html}
Revised by Dan Mirman to handle various corner cases
}
\examples{
WordLearnEx.gca <- code.poly(df=WordLearnEx, predictor="Block", poly.order=2)
Az.gca <- code.poly(df=Az, predictor="Time", poly.order=2, orthogonal=FALSE)
}

