% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{CP}
\alias{CP}
\title{Categorical perception of non-speech sounds}
\format{A data frame with 256 rows and 4 variables:
\describe{
 \item{Participant}{participant ID, N=32}
 \item{Type}{continuum type, Temporal or Spectral, between-participants}
 \item{Stimulus}{stimulus number along continuum, 1-8, within-participants}
 \item{d.prime}{auditory discrimination score, d'}
 }}
\source{
Based on Mirman, Holt, & McClelland (2004, J. Acoust. Soc. Am., 116:2, 1198-1207) \url{http://www.danmirman.org/pdfs/MirmanHoltMcClelland2004.pdf}
}
\usage{
CP
}
\description{
A dataset containing auditory discrimination data for non-speech sounds that vary on a temporal or spectral dimension.
}
\keyword{datasets}

