% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{DeviantBehavior}
\alias{DeviantBehavior}
\title{Tolerance of deviant behavior}
\format{A data frame with 80 rows and 5 variables
\describe{
 \item{SubjID}{participant ID, N=16}
 \item{Gender}{participant gender}
 \item{Exposure}{exposure to deviant behavior during adolescence (assumed time-invariant)}
 \item{Age}{age at reported tolerance}
 \item{Tolerance}{tolerance of deviant behavior, measured on 4-point scale (1 = very wrong, 2 = wrong, 3 = a little bit wrong, 4 = not wrong at all), averaged across items}
 }}
\source{
Table 11.1 (p. 220) in Willett (1997, "Measuring change: What individual growth modeling buys you." In Amsel, E. and Renninger, K. A. (Eds) Change and development: Issues of theory, method, and application, The Jean Piaget symposium series). \url{http://gseacademic.harvard.edu/~willetjo/change.htm}
}
\usage{
DeviantBehavior
}
\description{
Subset of National Youth Survey longitudinal data on tolerance for deviant behavior, exposure to deviant behavior, and gender. Deviant behavior was measured by a "9-item instrument that asked whether it was wrong for someone their age to do one of the following: cheat on tests, purposely destroy property of others, use marijuana, steal something worth less than $5, hit or threaten someone without reason, use alcohol, break into a building or vehicle to steal, sell hard drugs, or steal something worth more than $50 (Raudenbush & Chan, 1992)." (Willett, 1997, p. 219-220).
}
\keyword{datasets}

