% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{EducMH}
\alias{EducMH}
\title{School mental health services and educational achievement}
\format{A data frame with 1080 rows and 5 variables
\describe{
 \item{ID}{subject ID, factor, N=180}
 \item{Condition}{categorical variable with two between-subjects levels: Treatment (students who received mental health services) and Control (academically matched group of students who did not receive services)}
 \item{Year}{year of data collection, 2009-2014}
 \item{SDQ}{total difficulties score from Strengths and Difficulties Questionnaire, which is a brief behavioral screening for mental health. Only available for Treatment group. Lower scores indicate better mental health (less difficulties)}
 \item{Math}{score on standardized math test}
 }}
\usage{
EducMH
}
\description{
Invented longitudinal data for test of effect of school mental health services on educational achievement.
}
\keyword{datasets}

