% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get.pvalues}
\alias{get.pvalues}
\title{Get p-values for a multilevel regression model}
\usage{
get.pvalues(model, method = "normal")
}
\arguments{
\item{model}{a merMod model object fit by lmer}

\item{method}{p-value estimation method, defaults to normal, can also be KR or all}
}
\value{
data frame containing fixed effect estimates, SE, t-values, (degrees of freedom if appropriate) and p-values.
}
\description{
Takes a model fit with lmer and (optionally) an estimation method, returns a data frame containing fixed effect coefficient estimates, SE, t-values, and p-values. Can use normal approximation, Kenward-Roger approximation for degrees of freedom, or both. If Kenward-Roger approximation is used, result also contains approximated degrees of freedom.
}
\examples{
get.pvalues(m.full)
get.pvalues(m.full, "KR")
}

