% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{MotorLearning}
\alias{MotorLearning}
\title{Motor learning data}
\format{A data frame with 2400 rows and 5 variables:
\describe{
 \item{Subject}{subject ID, factor, N=20}
 \item{Difficulty}{difficulty of motor learning task, High or Low, within-participants}
 \item{Condition}{status of participant, Control or Impaired, within-participants}
 \item{Trial}{trial number, 1-30}
 \item{Accuracy}{proportion correct performance}
}}
\usage{
MotorLearning
}
\description{
Accuracy in a motor learning task as a function of trial number, task difficulty, and impairment. For example, playing Guitar Hero on easy vs. hard difficulty while sober vs. impaired (invented data).
}
\keyword{datasets}

