% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/num_params_ranef.R
\name{num_params_ranef}
\alias{num_params_ranef}
\title{Demonstrate number of random effects estimates}
\usage{
num_params_ranef(tim = 1:5, cond = 2:4)
}
\arguments{
\item{tim}{vector of number of (polynomial) time terms}

\item{cond}{vector of number of conditions}
}
\description{
Random effects for within-participant conditions can take two different
structures: a more flexible one with the condition factor on the left of
the pipe and a more constrained one with the condition factor on the right
of the pipe. This function makes a graph to demonstrate the cost of that
flexibility in terms of the number of random effect variances and
covariances to be estimated.
}
\details{
See Chapter 4 for more discussion of within-participant random effect structures.
}
\examples{
num_params_ranef()
}

