% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{risk.trial}
\alias{risk.item}
\alias{risk.subj}
\alias{risk.trial}
\title{Peer influence on adolescent risk-taking}
\format{A data frame with 2864 rows and 5 variables
\describe{
 \item{Item}{risk-taking scenario, can be in a peer context or no-peer context}
 \item{Context}{factor indicating context of risk-taking scenario: Peer (N=16) or NoPeer (N=14)}
 \item{Subject}{participant ID, N=120}
 \item{Resistance}{factor indicating each participant's resistance to peer influence level}
 \item{Score}{risk-taking score}
}}
\source{
Inspired by Chein et al. (2011, Developmental Science, 14, F1-F10)
}
\usage{
risk.trial

risk.item

risk.subj
}
\description{
Invented risk-taking data in peer and no-peer contexts for adolescents with high or low resistance to peer influence. In this hypothetical experiment, adolescents came to the laboratory and completed different risk-taking scenarios, which either involved a peer context or a no-peer context. (Note: some data are missing because not all participants completed all items). The participants also completed a standard measure of resistance to peer influence and those scores were used to divide participants into High and Low groups by median split.
}
\details{
Related data sets:
\itemize{
 \item risk.subj: subject means (averaged over items)
 \item risk.item: item (scenario) means (averaged over subjects)
 }
}
\keyword{datasets}

