% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{SJ}
\alias{SJ}
\title{Similarity judgments}
\format{A data frame with 810 rows and 5 variables
\describe{
 \item{Subject}{participant ID, N=45}
 \item{Item1}{first item in concept pair (N=18), each Item1 was presented with 3 different kinds of Item2}
 \item{Item2}{second item in concept pair, counterbalanced across participants}
 \item{SimRating}{each participant's similarity rating for concept pair}
 \item{Condition}{categorical variable with three levels: Distinct = concept pair shares at least one rare feature, Common = concept pair shares features that are common to many concepts, None = concept pair shares no features}
}}
\source{
Mirman & Magnuson (2009, Psychon. Bull. Rev., 16:4, 671-677) \url{http://www.danmirman.org/pdfs/MirmanMagnuson2009a.pdf}
}
\usage{
SJ
}
\description{
Semantic similarity judgments for concept pairs that share different kinds of features
}
\keyword{datasets}

