% Generated by roxygen2 (4.0.2): do not edit by hand
\name{t.df.demo}
\alias{t.df.demo}
\title{Compare normal distribution to t distributions with varying degrees of freedom}
\usage{
\method{t}{df.demo}(degf = c(1, 10, 100))
}
\arguments{
\item{degf}{vector of degrees of freedom for t distributions}
}
\description{
Display the Student's t distributions with various degrees of freedom and compare to the normal distribution
}
\examples{
t.df.demo()
t.df.demo(degf=c(1, 5, 15, 50, 500))
}

