% Generated by roxygen2 (4.0.2): do not edit by hand
\name{t_df_demo}
\alias{t_df_demo}
\title{Compare normal distribution to t distributions with varying degrees of freedom}
\usage{
t_df_demo(degf = c(1, 10, 100))
}
\arguments{
\item{degf}{vector of degrees of freedom for t distributions}
}
\description{
Display the Student's t distributions with various degrees of freedom and compare to the normal distribution
}
\examples{
t_df_demo()
t_df_demo(degf=c(1, 5, 15, 50, 500))
}

