% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{VisualSearchEx}
\alias{VisualSearchEx}
\title{Visual search response times}
\format{A data frame with 132 rows and 4 variables
\describe{
 \item{Participant}{participant ID, factor}
 \item{Dx}{diagnosis group: Aphasic (N=15) or Control (N=18)}
 \item{Set.Size}{number of objects on screen (including target): 1, 5, 15, or 30}
 \item{RT}{response time in ms from display onset}
 }}
\source{
LCDL Technical Report 2013.1: Preliminary validation of eye tracking and the "Visual World Paradigm"”" for participants with aphasia and limb apraxia. \url{http://www.danmirman.org/pdfs/LCDL-TR2013.01_AphasiaEyeTracking.pdf}
}
\usage{
VisualSearchEx
}
\description{
Conjunction visual search response times for participants with aphasia and matched neurologically-intact control participants as a function of number of objects in the display.
}
\keyword{datasets}

