% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{WeightMaintain3}
\alias{WeightMaintain3}
\title{Weight maintenance data}
\format{A data frame with 720 rows and 4 variables
\describe{
 \item{ID}{participant id, N=180}
 \item{Condition}{intervention type, factor with 3 levels: None, ED, MR}
 \item{Assessment}{weight change assessment relative to end of weight loss program and start of weight maintenance phase. 0 is baseline, 1-3 is years from start of maintenance}
 \item{WeightChange}{weight change in kg relative from start of maintenance phase}
}}
\source{
Based on Lowe et al. (2014, Obesity, 22, 94-100)
}
\usage{
WeightMaintain3
}
\description{
Invented data based on a real study. Overweight participants completed a 12-week weight loss program, then were randomly assigned to one of three weight maintenance conditions: (1) no intervention, (2) energy density (ED) = book and educational materials on purchasing and preparing foods lower in ED, (3) meal replacements (MR) = use MR to replace one meal and snack per day. Weight was assessed at baseline (start of maintenance), 12 months post, 24 months post, and 36 months post.
}
\keyword{datasets}

