% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{wisqars.suicide}
\alias{wisqars.suicide}
\title{Annual suicide rates by state}
\format{A data frame with 459 rows and 7 variables
\describe{
 \item{Region}{region for each state, factor with 4 levels: Northeast, South, Midwest, West (extracted from datasets::state)}
 \item{State}{state name, factor with 51 levels}
 \item{Year}{integer, 1999-2007}
 \item{Deaths}{total number of deaths by suicide for each state in each year}
 \item{Population}{each state's population for each year}
 \item{Crude.Rate}{suicide rate per 100,000 state residents}
 \item{Age.Adjusted.Rate}{suicide rate per 100,000 state residents adjusted for age distribution within each state (because suicide rates are not equally prevalent for all ages)}
}}
\source{
Web-based Injury Statistics Query and Reporting System (WISQARS) hosted by the Centers for Disease Control and Prevention: \url{http://www.cdc.gov/injury/wisqars/}
}
\usage{
wisqars.suicide
}
\description{
Annual suicide rates for each state (including DC) from 1999 to 2007.
}
\keyword{datasets}

